Multi-language support
Added multi languague support. It's a bit wonky and it needs better code but it works. For spotdl V4 there will be an organized systematic way of adding different languages to spotdl.

Added Spanish - Spain and English - United States as available languages. Will be working on docs on how to write the "Dictionary" file.